Dedups TestInMemoryStateInternals
